pyproject.toml
[tool.poetry]
name = "gemini-deep-research"
version = "0.1.0"
description = "CLI tool for Google Deep Research API"
authors = ["Filip Prochazka <dev@fprochazka.cz>"]
packages = [{include = "gemini_deep_research"}]

[tool.poetry.dependencies]
python = "^3.10"
typer = {extras = ["all"], version = "^0.15.0"}
rich = "^13.7.0"
python-dotenv = "^1.0.0"
google-genai = ">=1.56.0"

[tool.poetry.group.dev.dependencies]
ruff = "^0.14.0"
pytest = "^9.0.2"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.poetry.scripts]
gemini-deep-research = "gemini_deep_research.main:app"

[tool.ruff]
line-length = 120
target-version = "py310"

[tool.ruff.lint]
select = [
    "E",   # pycodestyle errors
    "W",   # pycodestyle warnings
    "F",   # pyflakes
    "I",   # isort
    "B",   # flake8-bugbear
    "C4",  # flake8-comprehensions
    "UP",  # pyupgrade
]
ignore = []

[tool.ruff.format]
quote-style = "double"
indent-style = "space"

[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = ["test_*.py"]
python_classes = ["Test*"]
python_functions = ["test_*"]
